Simplify and refactor how commands are read from the file
The first line is checked off the bat, which means we don't need the read state variable to tell us how to interpret each line because all proceeding lines are parsed as commands.

Command parsing for the line was moved to `Command.parseLeaky`. This keeps the surface area for additional commands to mostly the enum `CommandType` and `Command`